Added the ability to create a prenote file
This commit adds the ability to create a prenote file by using $0.00 as the transaction amount for each entry.

A prenote is a zero-dolloar transaction that allows you to test your ACH capabilities with a certain account. Some accounts block ACH. By sending a prenote file you can test that account. 

***Keep in mind that by sending a prenote, you will likely lock that account from other transactions for about a week.